styling complete for captive-portal page
