+++ client/src/App.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 import { Box } from "@mui/material"
-import ClusterViewContainer from "./features/cluster-view/containers/ClusterViewContainer"
+
 import MiniDrawer from "./features/mini-drawer/MiniDrawer"
 
 const App = () => {
