+++ client/src/features/captive-portal/containers/CaptivePortal.js
@@ -1,0 +1,12 @@
+import { jsx as _jsx, jsxs as _jsxs } from "react/jsx-runtime";
+import Stack from '@mui/material/Stack';
+import CaptiveForm from "../components/CaptiveForm";
+import CaptiveFormContent from "../components/CaptiveFormContent";
+export default function CaptivePortal() {
+    return (_jsxs(Stack, { direction: { xs: 'column-reverse', md: 'row' }, sx: {
+            display: 'flex',
+            justifyContent: 'center',
+            alignItems: 'center',
+            height: '100vh', // Full viewport height
+        }, children: [_jsx(CaptiveFormContent, {}), _jsx(CaptiveForm, {})] }));
+}
