+++ client/src/main.js
@@ -7,7 +7,7 @@ import LandingPage from "./features/landing-page/LandingPage";
 import { store } from "./app/store";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import "./index.css";
-import CaptivePortal from "./features/captive-portal/CaptivePortal";
+import CaptivePortal from "./features/captive-portal/containers/CaptivePortal";
 import LogPage from "./features/log-page/LogPage";
 // import material UI fonts
 import "@fontsource/roboto/300.css";
